// src/types.ts
export interface SimpleOptions {
  url: string;
  title: string;
  x: number;
  y: number;
  value: number;
  title2: string;
  x2: number;
  y2: number;
  value2: number;
  title3: string;
  x3: number;
  y3: number;
  value3: number;
  title4: string;
  x4: number;
  y4: number;
  value4: number;
  title5: string;
  x5: number;
  y5: number;
  value5: number;
  monitorAmount: number;
}

export const defaults: SimpleOptions = {
  url: '',
  title: '',
  x: 0,
  y: 0,
  title2: '',
  x2: 0,
  y2: 0,
  title3: '',
  x3: 0,
  y3: 0,
  title4: '',
  x4: 0,
  y4: 0,
  title5: '',
  x5: 0,
  y5: 0,
  value: 0,
  value2: 0,
  value3: 0,
  value4: 0,
  value5: 0,
  monitorAmount: 0
};